+++ packages/article-summary/src/__tests__/shared.js
@@ -5,6 +5,7 @@ import ArticleSummary, {
   summarise
 } from "../article-summary";
 import defaultFixture from "../../fixtures/default";
+import opinionBylineFixture from "../../fixtures/opinion-byline";
 import articleMultiFixture from "../../fixtures/article-multi";
 import emptyParagraphFixture from "../../fixtures/article-empty-paragraph";
 import noBylineFixture from "../../fixtures/no-byline";
@@ -33,6 +34,14 @@ export default () => {
   it("renders an article-summary component with a single paragraph", () => {
     const tree = renderer
       .create(<ArticleSummary {...defaultFixture} />)
+      .toJSON();
+
+    expect(tree).toMatchSnapshot();
+  });
+
+  it("renders an article-summary with opinion byline", () => {
+    const tree = renderer
+      .create(<ArticleSummary {...opinionBylineFixture} />)
       .toJSON();
 
     expect(tree).toMatchSnapshot();