+++ packages/article/src/__tests__/web/article.web.test.js
@@ -60,6 +60,8 @@ describe("Article test on web", () => {
           {...articleFixtureNoLeadAsset.data}
           analyticsStream={() => {}}
           adConfig={adConfig}
+          onRelatedArticlePress={() => {}}
+          onAuthorPress={() => {}}
         />
       )
       .toJSON();