+++ packages/slice/src/templates/leadandtwo/config.web.js
@@ -1,5 +1,6 @@
 import { View } from "react-native";
 import withResponsiveStyles from "@times-components/responsive-styles";
+import { spacing } from "@times-components/styleguide";
 
 const leadSummaryConfig = {
   1: [125],
@@ -11,7 +12,6 @@ export const getLeadConfig = ({ itemCount }) => ({
   contentContainerClass: "leadContentContainerClass",
   headlineClass: "leadHeadlineClass",
   imageContainerClass: "leadImageContainerClass",
-  summaryClass: "leadSummaryClass",
   summaryConfig: {
     lengths: leadSummaryConfig[itemCount],
     type: "lead"
@@ -20,7 +20,6 @@ export const getLeadConfig = ({ itemCount }) => ({
 
 export const getSupportConfig = () => ({
   contentContainerClass: "supportContentContainerClass",
-  headlineClass: "supportHeadlineClass",
   imageContainerClass: "supportImageContainerClass",
   summaryClass: "supportSummaryClass"
 });
@@ -43,12 +42,12 @@ export const getConfigWrapper = ({ supportCount }) => {
     `,
     mediumUp: () => {
       const withSupportsImageStyle = `
-        margin-bottom: 10px;
+        margin-bottom: ${spacing(2)};
         min-width: 370px;
       `;
 
       const withoutSupportsImageStyle = `
-        padding-right: 10px;
+        padding-right: ${spacing(2)};
       `;
 
       return `
@@ -56,7 +55,6 @@ export const getConfigWrapper = ({ supportCount }) => {
           font-size: 30px;
           line-height: 30px;
         }
-
         .leadImageContainerClass {
           flex: 1;
           margin-bottom: 0;
@@ -69,19 +67,17 @@ export const getConfigWrapper = ({ supportCount }) => {
         }
 
         .leadContentContainerClass {
-          flex-grow: 1;
           flex-basis: 0 !important;
+          flex-grow: 1;
+          margin-bottom: ${spacing(2)};
           min-width: 300px;
         }
-
         .supportImageContainerClass {
           display: block;
         }
-
         .leadSummary125Class {
           display: ${supportCount === 2 ? "none" : "block"};
         }
-
         .leadSummary175Class {
           display: ${supportCount === 2 ? "block" : "none"};
         }
@@ -97,24 +93,23 @@ export const getConfigWrapper = ({ supportCount }) => {
         margin-bottom: 0;
         max-width: 185px;
         min-width: auto;
-        padding-right: 10px;
+        padding-right: ${spacing(2)};
       `;
 
       const twoSupportContentStyle = `
         flex: 1;
+        margin-bottom: 0;
         min-width: 100px;
       `;
 
       return `
         .leadImageContainerClass {
-          padding-right: 10px;
+          padding-right: ${spacing(2)};
           ${supportCount === 2 ? twoLeadImageStyle : ``}
         }
-
         .supportImageContainerClass {
           ${supportCount === 2 ? twoSupportImageStyle : ``}
         }
-
         .supportContentContainerClass {
           ${supportCount === 2 ? twoSupportContentStyle : ``}
         }