+++ packages/slice/src/templates/leadandtwo/index.js
@@ -2,7 +2,6 @@ import React from "react";
 import { View } from "react-native";
 import styles from "../styles";
 import { propTypes, defaultProps } from "./proptypes";
-import leadAndTwoStyles from "./styles";
 import { leadConfig, supportConfig } from "./config";
 
 const LeadAndTwoSlice = ({ lead, renderSupports }) => (
@@ -12,7 +11,7 @@ const LeadAndTwoSlice = ({ lead, renderSupports }) => (
     </View>
     {renderSupports(supportConfig).map(support => (
       <View key={support.props.id} style={styles.itemContainer}>
-        <View style={[styles.item, leadAndTwoStyles.supports]}>{support}</View>
+        <View style={styles.item}>{support}</View>
       </View>
     ))}
   </View>