+++ packages/slice/src/templates/opinionandtwo/responsive.web.js
@@ -2,65 +2,25 @@ import { View } from "react-native";
 import withResponsiveStyles, {
   config
 } from "@times-components/responsive-styles";
-import { colours } from "@times-components/styleguide";
+import { colours, spacing } from "@times-components/styleguide";
 
-export const getSeparator = ({ itemCount }) => {
-  const Separator = withResponsiveStyles(View, {
-    base: () => `
-      border-bottom-style: solid;
-      border-bottom-width: 1px;
-      border-bottom-color: ${colours.functional.keyline};
-      flex: 1;
-      margin-bottom: 10px;
-      margin-top: 10px;
-      min-width: auto;
-    `,
-    mediumUp: () => `
-      border-bottom: none;  
-      border-right-style: solid;
-      border-right-width: 1px;
-      border-right-color: ${colours.functional.keyline};
-      display: ${itemCount === 3 ? "none" : "block"};
-      flex: 0 !important;
-      margin: 0;
-    `,
-    wideUp: () => `
-      display: block;
-    `
-  });
-  Separator.displayName = "Separator";
-  return Separator;
-};
-
-export const getContainer = ({ hasSupports }) => {
+export const getContainer = ({ supportCount }) => {
   const Container = withResponsiveStyles(View, {
     base: () => `
       flex: 1;
       flex-direction: column;
       flex-wrap: wrap;
-      padding-bottom: 10px;
-      padding-top: 10px;
       height: auto;
+      padding-bottom: ${spacing(2)};
+      padding-top: ${spacing(2)};
       width: 100%;
     `,
-    mediumUp: () => {
-      const withoutSupportsStyle = `
-        padding-left: 0px;
-        padding-right: 0px;
-        width: ${config.mediumBpWidth};
-      `;
-
-      const withSupportsStyle = `
-        width: 100%;
-      `;
-
-      return `
-        flex-direction: row;
-        ${hasSupports ? withSupportsStyle : withoutSupportsStyle}
-      `;
-    },
+    mediumUp: () => `
+      flex-direction: row;
+      width: ${supportCount === 1 ? "100%" : config.mediumBpWidth};
+    `,
     wideUp: () => `
-      width: ${hasSupports ? "100%" : config.wideBpWidth};
+      width: ${supportCount === 0 ? config.wideBpWidth : "100%"};
     `
   });
   Container.displayName = "Container";
@@ -72,8 +32,8 @@ export const getOpinionContainer = ({ hasSupports, supportCount }) => {
     base: () => `
       flex: 1;
       flex-grow: 1;
-      padding-left: 10px;
-      padding-right: 10px;
+      padding-left: ${spacing(2)};
+      padding-right: ${spacing(2)};
       width: ${hasSupports ? "auto" : "100%"};
     `,
     mediumUp: () => {
@@ -84,40 +44,64 @@ export const getOpinionContainer = ({ hasSupports, supportCount }) => {
       `;
       const withSupportsStyle = `
         flex-basis: 0 !important;
-        flex-grow: ${supportCount === 1 ? "3" : "2"};
-        padding-left: 10px;
-        padding-right: 10px;
+        flex-grow: ${supportCount === 1 ? "3" : "1"};
+        padding-left: ${spacing(2)};
+        padding-right: ${spacing(2)};
       `;
       const twoSupportStyle = `
+        border-bottom-color: ${colours.functional.keyline};
         border-bottom-style: solid;
         border-bottom-width: 1px;
-        border-bottom-color: ${colours.functional.keyline};
-        padding-bottom: 10px;
-        margin-bottom: 10px;
+        margin-bottom: ${spacing(2)};
         min-width: 100%;
+        padding-bottom: ${spacing(2)};
+        padding-left: 0;
+        padding-right: 0;
       `;
       return `
         ${hasSupports ? withSupportsStyle : withoutSupportsStyle}
         ${supportCount === 2 ? twoSupportStyle : ``}
       `;
     },
-    wideUp: () => {
-      const getFlexGrow = () => {
-        if (supportCount === 0) return "1.5";
-        if (supportCount === 1) return "2.75";
-        return "2";
-      };
-      return `
-        border-bottom: none;
-        flex-grow: ${getFlexGrow()};
-        padding-bottom: 0;
-        margin-bottom: 0;
-        min-width: auto;
-      `;
-    }
+    wideUp: () => `
+      border-bottom: none;
+      margin-bottom: 0;
+      min-width: auto;
+      padding-bottom: 0;
+      padding-left: ${spacing(2)};
+      padding-right: ${spacing(2)};
+    `
   });
   OpinionContainer.displayName = "OpinionContainer";
   return OpinionContainer;
+};
+
+export const getSeparator = ({ itemCount }) => {
+  const Separator = withResponsiveStyles(View, {
+    base: () => `
+      border-bottom-color: ${colours.functional.keyline};
+      border-bottom-style: solid;
+      border-bottom-width: 1px;
+      flex: 1;
+      margin-bottom: ${spacing(2)};
+      margin-top: ${spacing(2)};
+      min-width: auto;
+    `,
+    mediumUp: () => `
+      border-bottom: none;  
+      border-right-color: ${colours.functional.keyline};
+      border-right-style: solid;
+      border-right-width: 1px;
+      display: ${itemCount === 3 ? "none" : "block"};
+      flex: 0 !important;
+      margin: 0;
+    `,
+    wideUp: () => `
+      display: block;
+    `
+  });
+  Separator.displayName = "Separator";
+  return Separator;
 };
 
 export const getSupportsContainer = ({ supportCount }) => {
@@ -138,8 +122,7 @@ export const getSupportsContainer = ({ supportCount }) => {
       `;
     },
     wideUp: () => `
-      flex-grow: ${supportCount === 2 ? "2" : "1"};
-      min-width: auto;
+      min-width: ${supportCount === 1 ? "400px" : "auto"};
     `
   });
   SupportsContainer.displayName = "SupportsContainer";
@@ -150,45 +133,59 @@ export const getSupportContainer = ({ index, supportCount }) => {
   const SupportContainer = withResponsiveStyles(View, {
     base: () => {
       const secondSupportStyle = `
+        border-top-color: ${colours.functional.keyline};
         border-top-style: solid;
         border-top-width: 1px;
-        border-top-color: ${colours.functional.keyline};
-        margin-top: 10px;
-        padding-top: 10px;
+        margin-top: ${spacing(2)};
+        padding-top: ${spacing(2)};
       `;
 
       return `
         flex: 1;
         flex-wrap: wrap;
         min-height: auto;
-        padding-left: 10px;
-        padding-right: 10px;
+        padding-left: ${spacing(2)};
+        padding-right: ${spacing(2)};
         ${index === 1 ? secondSupportStyle : ``}
       `;
     },
     mediumUp: () => {
+      const firstSupportStyle = `
+        padding-left: ${supportCount === 2 ? `0` : `${spacing(2)}`};
+      `;
+
       const secondSupportStyle = `
+        border-left-color: ${colours.functional.keyline};
         border-left-style: solid;
         border-left-width: 1px;
-        border-left-color: ${colours.functional.keyline};
         border-top: none;
-        margin-left: 10px;
+        margin-left: ${spacing(2)};
         margin-top: 0;
-        padding-left: 10px;
-        padding-right: 10px;
+        padding-right: ${supportCount === 2 ? `0` : `${spacing(2)}`};
         padding-top: 0;
       `;
       const oneSupportStyle = `
         max-width: 100%;
-        padding-right: 10px;
+        padding-right: ${spacing(2)};
       `;
       const twoSupportStyle = `
         max-width: 50%;
       `;
       return `
+        flex-basis: 0 !important;
         padding-right: 0;
-        ${index === 1 ? secondSupportStyle : ``}
+        ${index === 1 ? secondSupportStyle : firstSupportStyle}
         ${supportCount === 2 ? twoSupportStyle : oneSupportStyle}
+      `;
+    },
+    wideUp: () => {
+      const paddingRightStyle = `
+        padding-right: ${spacing(2)};
+      `;
+      return `
+        padding-left: ${spacing(2)};
+        ${index === 1 ? paddingRightStyle : ``}
+        ${supportCount === 1 ? paddingRightStyle : ``}
       `;
     }
   });