+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.14"></a>
+## [0.3.14](https://github.com/newsuk/times-components/compare/@times-components/icons@0.3.13...@times-components/icons@0.3.14) (2018-03-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/icons
+
 <a name="0.3.13"></a>
 ## [0.3.13](https://github.com/newsuk/times-components/compare/@times-components/icons@0.3.12...@times-components/icons@0.3.13) (2018-03-28)