+++ packages/video-label/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.1.18"></a>
+## [0.1.18](https://github.com/newsuk/times-components/compare/@times-components/video-label@0.1.17...@times-components/video-label@0.1.18) (2018-03-29)
+
+
+
+
+**Note:** Version bump only for package @times-components/video-label
+
 <a name="0.1.17"></a>
 ## [0.1.17](https://github.com/newsuk/times-components/compare/@times-components/video-label@0.1.16...@times-components/video-label@0.1.17) (2018-03-28)